multiple fixes

+++ src/graph.rs
@@ -936,11 +936,11 @@ pub struct Target {
     pub links: Box<[u64]>, // the bandwidth of each link
     pub paths: Box<[Box<[usize]>]>, // the i*n+j element is the links that i->j uses (currently only one path between each pair)
     pub sinks: Box<[String]>, // sink nodes
-    pub nccls: BTreeMap<String, [f64; 4]> // the key is a comma separated sorted list of device names, the values are [coef1, interc1, coef2, interc2]. The model is time = exp(max( coef1 * log2(size) + interc1, coef2 * log2(size) + interc2 )). The size unit is KB.
+    pub nccls: BTreeMap<String, [i64; 4]> // the key is a comma separated sorted list of device names, the values are [coef1, interc1, coef2, interc2]. The model is time = max( coef1 * size + interc1, coef2 * size + interc2 ). The size unit is KB.
 }
 
 impl Target {
-    pub fn new(pb: GraphDef, devices: Box<[String]>, links: Box<[u64]>, paths: Box<[Box<[usize]>]>, sinks: Box<[String]>, nccls: BTreeMap<String, [f64; 4]>) -> Self {
+    pub fn new(pb: GraphDef, devices: Box<[String]>, links: Box<[u64]>, paths: Box<[Box<[usize]>]>, sinks: Box<[String]>, nccls: BTreeMap<String, [i64; 4]>) -> Self {
         Target { pb, devices, links, paths, sinks, nccls }
     }
